\name{col.norm}
\alias{col.norm}

\title{
Compute the L2 norm of each column
}

\description{
The function returns a vector containing the L2 norm of each column for a given matrix.
}

\usage{
col.norm(X)
}

\arguments{
  \item{X}{
a matrix for which the column norms are computed.
}
}

\value{
A vector containing the L2 norm of the columns of X is returned.
}

\author{
Haeran Cho
}
